research/multi_representation_adversary/requirements.txt
absl-py
git+git://github.com/cleverhans-lab/cleverhans@v3.1.0
gin-config
neural-structured-learning>=1.3.0
numpy~=1.19.2
tensorflow~=2.4.1
tensorflow-addons
tensorflow-datasets>=4.0.0
